test: split language tests and add extra logs
Split language tests so they can be run seperately and also so we can
pinpoint which specific words pass/fail.

Also print the full, indented JSON if there's a difference between the
expected and actual result so it's easier to debug and possibly update
the testOutput.json if need be.

Add parameterized as new dependency.